Some more refactoring to enable the metadata retrieval for the SQL Server platform as well. Bean table filter and enum table filter are now already applied when retrieving the metadata and no longer in the code generators.

+++ generator/src/main/java/org/noorm/generator/beangenerator/BeanClassDescriptor.java
@@ -21,7 +21,7 @@ public class BeanClassDescriptor {
     private String[] primaryKeyColumnNames;
     private String[] primaryKeyJavaNames;
 	private String sequenceName;
-    private Long sequenceIncrement;
+    private Integer sequenceIncrement;
     private boolean useInlineSequenceValueGeneration = false;
     private String versionColumnName;
     private String versionColumnJavaName = "";
@@ -108,11 +108,11 @@ public class BeanClassDescriptor {
 		sequenceName = pSequenceName;
 	}
 
-    public Long getSequenceIncrement() {
+    public Integer getSequenceIncrement() {
         return sequenceIncrement;
     }
 
-    public void setSequenceIncrement(final Long pSequenceIncrement) {
+    public void setSequenceIncrement(final Integer pSequenceIncrement) {
         sequenceIncrement = pSequenceIncrement;
     }
